# api/openApi.yaml
openapi: 3.0.3
info:
  title: Auth API
  version: "1.0.0"
  description: |
    A comprehensive authentication and authorization API for the GroundGo library.
    Provides user management, role-based access control, and hierarchical path-specific claims.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: 'http://127.0.0.1:8080/api/v1'
    description: Development server
  - url: https://api.groundgo.com/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      operationId:  postAuthRefresh
      security: 
        - RefreshToken: []
      tags:
        - Authentication
      summary: Refresh access token
      description: Generate new access token using refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/InvalidRefreshToken'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{id}:
    get:
      operationId: getUserByID
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: "#/components/responses/Forbidden"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  
    patch: 
      operationId: updateUserByID
      tags: 
        -  Users
      summary: Update a single user by ID
      description: |
        Update a single user. For use by admin.
        - Partial updates are allowed
        - To update your own profile if you you do not have permission use path /users/me
      parameters: 
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses: 
        '200':
          description: User updated successfully
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: "#/components/responses/Forbidden"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      
    delete: 
      operationId: deleteUserByID
      tags: 
        -  Users
      summary: Permanently delete a single user by ID
      parameters: 
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        '204':
          description: "delete Successful"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: "#/components/responses/Forbidden"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/me:
    get:
      operationId: getOwnUser
      tags:
        - Self Service
      summary: Get logged in user
      description: Retrieve a users own User record
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: "#/components/responses/Forbidden"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    RefreshToken:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token stored in httpOnly cookie
  responses:
    BadRequest:
      description: Invalid request format or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidJson:
              summary: Invalid JSON format
              value:
                error: "invalid JSON format"
                details: "Expected valid JSON object"
            missingField:
              summary: Missing required field
              value:
                error: "validation failed"
                details: "email is required"
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing authentication token
              value:
                error: "authentication required"
                details: "Bearer token must be provided"
            invalidToken:
              summary: Invalid token
              value:
                error: "invalid token"
                details: "token is expired or malformed"
    Forbidden:
      description: Access denied - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "access denied"
            details: "insufficient permissions for this operation"
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "resource not found"
            details: "user with specified ID does not exist"
...
    InvalidRefreshToken:
      description: Invalid or expired refresh token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing refresh token
              value:
                error: "missing refresh token"
                details: "Refresh token cookie not found"
            invalidToken:
              summary: Invalid refresh token
              value:
                error: "invalid refresh token"
                details: "token is expired, invalid, or has been revoked"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "internal server error"
            details: "an unexpected error occurred"
  parameters:
    UserIdParam:
      name: id
      in: path
      required: true
      description: User ID
      schema:
        type: string
        format: uuid
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
          nullable: true
      required:
        - error
...
    UserUpdateRequest:
      type: object
      properties:
        email:
          $ref: "#/components/schemas/Email"
        role:
          $ref: "#/components/schemas/Role"
        claims:
          $ref: '#/components/schemas/Claims'
        oauthProvider:
          type: string
          nullable: true
          description: OAuth provider (if OAuth login used)
        isActive:
          type: boolean
          description: Whether the user account is active
      additionalProperties: false
      description: Fields allowed to be updated in a user record
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          $ref: "#/components/schemas/Email"
        role:
          $ref: "#/components/schemas/Role"
        claims:
          $ref: '#/components/schemas/Claims'
        oauthProvider:
          type: string
          nullable: true
          description: OAuth provider (if OAuth login used)
        oauthId:
          type: string
          nullable: true
          description: OAuth provider user ID
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        isActive:
          type: boolean
          description: Whether the user account is active
      required:
        - id
        - email
        - role
        - claims
        - createdAt
        - updatedAt
        - isActive
    Email:
      type: string
      format: email
      description: User's email address
      example: "user@example.com"
    Claims:
      type: object
      description: |
       Claims is an object mapping API paths to user roles.

        - Each key is an API path (e.g. "/", "/users")
        - Each value is the role required to access that path
        - Roles are defined in the Role schema (see #/components/schemas/Role)
        - The root path "/" and the users assigned role will always be synced up. Updating the root path will update the users role and vice versa.
        - If both the root path and the role are added but different, the role will take precident.
      additionalProperties:
        $ref: "#/components/schemas/Role"
      example:
        "/": Admin
        "/users": User
    Role:
      type: string
      enum:
        - guest
        - readonly
        - user
        - auditor
        - editor
        - moderator
        - support
        - admin
        - owner
        - sysadmin
      description: Possible user roles in the system

    # Response schemas
    TokenResponse:
      type: object
      properties:
        token:
          type: string
          description: New JWT access token
        expiresIn:
          type: integer
          description: Token expiration time in seconds
      required:
        - token
        - expiresIn
...
    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
        - user
...